+++ src/Model/SettingValue.php
@@ -3,8 +3,8 @@
  * Settings Manager
  *
  * @license http://opensource.org/licenses/MIT
- * @link https://github.com/caseyamcl/settings_manager
- * @package caseyamcl/settings_manager
+ * @link https://github.com/caseyamcl/settings-manager
+ * @package caseyamcl/settings-manager
  * @author Casey McLaughlin <caseyamcl@gmail.com>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -19,7 +19,9 @@ use SettingsManager\Contract\SettingsProviderInterface;
 use SettingsManager\Contract\SettingValueInterface;
 
 /**
- * Class SettingValue
+ * Setting value - Stores a value for a given setting
+ *
+ * @author Casey McLaughlin <caseyamcl@gmail.com>
  */
 class SettingValue implements SettingValueInterface
 {