+++ src/Provider/CascadingSettingProvider.php
@@ -3,8 +3,8 @@
  * Settings Manager
  *
  * @license http://opensource.org/licenses/MIT
- * @link https://github.com/caseyamcl/settings_manager
- * @package caseyamcl/settings_manager
+ * @link https://github.com/caseyamcl/settings-manager
+ * @package caseyamcl/settings-manager
  * @author Casey McLaughlin <caseyamcl@gmail.com>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -15,15 +15,22 @@
 
 namespace SettingsManager\Provider;
 
+use SettingsManager\Behavior\SettingProviderTrait;
 use SettingsManager\Contract\SettingsProviderInterface;
 use SettingsManager\Contract\SettingValueInterface;
 use SettingsManager\Exception\ImmutableSettingOverrideException;
 
 /**
- * Class CascadeProvider
+ * Cascading setting provider
+ *
+ * Reads settings from multiple providers
+ *
+ * @author Casey McLaughlin <caseyamcl@gmail.com>
  */
 class CascadingSettingProvider implements SettingsProviderInterface
 {
+    use SettingProviderTrait;
+
     /**
      * @var array|SettingsProviderInterface[]
      */
@@ -95,15 +102,5 @@ class CascadingSettingProvider implements SettingsProviderInterface
     public function findSettingValue(string $name): ?SettingValueInterface
     {
         return $this->getSettingValues()[$name] ?? null;
-    }
-
-
-    public function findValue(string $settingName)
-    {
-        if ($setting = $this->findSettingValue($settingName)) {
-            return $setting->getValue();
-        } else {
-            return null;
-        }
     }
 }